chatbot_backend/requirements.txt
flask
flask-cors
openai
python-dotenv
ollama
flask_sqlalchemy
PyJWT
psycopg2-binary